requirements.txt
asgiref==3.5.2
autopep8==2.0.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.50.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
idna==3.4
numpy==1.23.5
oauthlib==3.2.2
pandas==1.5.2
psycopg2==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyforbes==1.0.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tweepy==4.12.1
tzdata==2022.7
urllib3==1.26.13
uvicorn==0.20.0
wikipedia==1.4.0
